move query options to ir module
