+++ fluree-db-query/tests/groupby_aggregate_tests.rs
@@ -14,7 +14,7 @@ use fluree_db_query::execute::{execute, ContextConfig, ExecutableQuery};
 use fluree_db_query::groupby::GroupByOperator;
 use fluree_db_query::ir::{Expression, FilterValue, Pattern};
 use fluree_db_query::operator::Operator;
-use fluree_db_query::options::QueryOptions;
+use fluree_db_query::ir::QueryOptions;
 use fluree_db_query::ir::{Query, QueryOutput};
 use fluree_db_query::var_registry::{VarId, VarRegistry};
 use fluree_graph_json_ld::ParsedContext;
