+++ fluree-db-query/tests/owl2rl_property_rules_integration_tests.rs
@@ -15,9 +15,9 @@ use fluree_db_core::value::FlakeValue;
 use fluree_db_core::{GraphDbRef, LedgerSnapshot, Sid};
 use fluree_db_query::binding::{Binding, RowAccess};
 use fluree_db_query::execute::{execute, ContextConfig, ExecutableQuery};
-use fluree_db_query::options::QueryOptions;
+use fluree_db_query::ir::QueryOptions;
 use fluree_db_query::parse::{parse_query, MemoryEncoder};
-use fluree_db_query::rewrite::ReasoningModes;
+use fluree_db_query::ir::ReasoningModes;
 use fluree_db_query::var_registry::VarRegistry;
 use fluree_db_reasoner::{reason_owl2rl, ReasoningCache, ReasoningOptions};
 use serde_json::json;
