+++ fluree-db-sparql/src/lower/ask.rs
@@ -5,7 +5,7 @@
 
 use crate::ast::query::AskQuery;
 
-use fluree_db_query::options::QueryOptions;
+use fluree_db_query::ir::QueryOptions;
 use fluree_db_query::parse::encode::IriEncoder;
 use fluree_db_query::ir::{Query, QueryOutput};
