+++ fluree-db-sparql/src/lower/select.rs
@@ -13,7 +13,7 @@ use crate::span::SourceSpan;
 
 use fluree_db_query::aggregate::AggregateSpec;
 use fluree_db_query::ir::{Expression, FilterValue, Pattern, SubqueryPattern};
-use fluree_db_query::options::QueryOptions;
+use fluree_db_query::ir::QueryOptions;
 use fluree_db_query::parse::encode::IriEncoder;
 use fluree_db_query::sort::{SortDirection, SortSpec};
 use fluree_db_query::var_registry::VarId;
